requirements.txt
discord.py
bs4
requests
slpp
python-dotenv
reactionmenu